Match `starts_with` and `ends_with` functions with Python implementation
